fix: install vscode extensions specified in setup bundles
The bug occured because `VSCODE_EXTENSIONS` used to list the vscode
extensions for installation is a local variable. When switching to the
standard user with `sudo`, it got reinitialized with the list of
standard extensions. To fix the bug, the installation command is made
an environment variable and preserved while switching the user.

fixes: #10