refactor: preserve vscode extensions list when switching to standard user (#15)
This change reworks the previous bugfix #10.

To fix the issue of vscode extensions list being reinitialized when
switching to the standard user, the previous bugfix instead computed
the install command as the root user and preserved that as an
environment variable when switching to the standard user.

However, this creates a limitation that the vscode extensions list
cannot be easily extended by setup bundles that run only as the
standard user. Hence, the bugfix needs to be reworked such that the
list of vscode extensions itself is preserved when switching.

Since bash cannot natively store an array as an environment variable,
the standard workaround of serializing it to a string, then
deserializing it back to an array is used with this refactoring. In
addition, this change initializes the `VSCODE_EXTENSIONS` variable while
the script runs as the root user, thus preventing reinitialization
when it switches to the standard user.

fixes: #14