feat: Describe `whistle.bash` setup arguments in `whistle.ps1` help (#26)
Refactor the help text in `whistle.bash` into a separate text file that
is now reused by both `whistle.bash` and `whistle.ps1`. This enables
the users to access the help text directly from the Powershell script.

It also addresses another problem wherein even if the user specified
`-h` argument to `-SetupArgs` as the current documentation suggests, it
requires creating an entire new WSL2 instance to run the bash script
to see the help text!

fixes: #24